Make "Download" button go to gnunet.git
The download button was leading to the page with all of the repos under
the GNUnet umbrella. For a download of GNUnet, it makes sense to go
directly to the GNUnet core git. I also added an indication that this
download button leads to a git repo, as is untypical for a "Download"
link.