tutorials: add section on configuring bandwidth
Increasing bandwidth from the default has been an overlooked step in
getting setup with a working GNUnet node. The basics on how to do this
are now added to the tutorials, though I think it could use some more
information on the valid formats for values, which I am unsure of at
this time. eg. is 10 GiB a valid value for WAN_QUOTA_IN?