Alexandros Theodotou <alex@zrythm.org> writes:
Re: our discussion on IRC:

There are at least two files in the GNUnet website source code that
have unclear licenses, rendering them proprietary.
https://git.gnunet.org/www.git/tree/i18nfix.py
https://git.gnunet.org/www.git/tree/template.py

These files have no license or copyright notices. They simply say that
"this file is in the public domain". In at least European countries,
saying that "this file is in the public domain" is not enough to put
something in the public domain. You have to explicitly use a license
such as CC0 or WTFPL, otherwise the files are "all rights reserved".

I'm not really sure you can assume that these files are GPL'd either,
otherwise they'd have a copyright notice. So in effect, these files are
proprietary. I can't copy them and use them in my (A)GPL project.

For such files, if you want to be lax, the GNU package maintainer
guidelines generally recommend this:
"""
Copyright (C) 2019 <name>

Copying and distribution of this file, with or without modification,
are permitted in any medium without royalty provided the copyright
notice and this notice are preserved.  This file is offered as-is,
without any warranty.
"""

Can we add appropriate headers or footers to the files please?